--- pkt/__version__.py
@@ -1,3 +1,0 @@
-"""Current version of package PheKnowLator"""
-
-__version__ = "2.0.0"
